backend/requirements(1).txt
Flask
Flask-CORS
Flask-SQLAlchemy
SQLAlchemy
python-dotenv
pandas
openpyxl
xlrd
requests
Werkzeug
reportlab
gunicorn
gradio_client
python-dotenv